Changed `indented` to true
I am not sure if this will have other negative effects, but I
really dislike seeing the bullets in the default layout to the left
of the other text margins.  The list items should be intented under
the title with the bullets aligned with the edge of the paragraph text.